Bugs and stylings

+++ trainmate/src/api/ExerciseApi.ts
@@ -106,7 +106,7 @@ export const getExerciseFromCategory = async (category_id: String) => {
     }
   };
 
-export const editExercise = async (exerciseData: { name: string, calories_per_hour: number | string }, exercise_id: string) => {
+export const editExercise = async (exerciseData: { name: string, calories_per_hour: number | string, training_muscle: string }, exercise_id: string) => {
         
         const token = getAuthToken();
         if (!token) throw new Error('Token no encontrado');
@@ -157,48 +157,56 @@ export const editExercise = async (exerciseData: { name: string, calories_per_ho
         }
     }
 
-export const deleteExercise = async (exercise_id: string) => {
-    const token = getAuthToken();
-    if (!token) throw new Error('Token no encontrado');
+export const deleteExercise = async (exercise_id: string, trainings: any) => {
+    // Primero hago validacion de que el ejercicio no este en ningun entrenamiento
+    const exerciseInTraining = trainings.find((training: any) => training.exercises.find((ex: any) => ex.exercise_id === exercise_id));
+    const exerciseInTrainingNew = trainings.find((training: any) => training.exercises.find((ex: any) => ex.id === exercise_id)); // Esto es temporal, pq si se acaba de crear un training, el exercise no tiene el exercise_id sino un id, hay que corrrgir eso
+    if (exerciseInTraining || exerciseInTrainingNew) {
+        throw new Error('Exercise is in use in a training, it cannot be deleted');
+    }
+    else {
+      const token = getAuthToken();
+        if (!token) throw new Error('Token no encontrado');
 
-    try {
-    const response = await fetch(`${BASE_URL}/api/exercise/delete-exercise/${exercise_id}`, {
-        method: 'DELETE',
-        headers: {
-        'Authorization': token,
-        },
-    });
-
-    // Si la respuesta es 401, intentamos renovar el token
-    if (response.status === 403 || response.status === 401) {
-        console.log('Token expirado, intentando renovar...');
-        const newToken = await refreshAuthToken(); // Renueva el token
-        // Intentamos la solicitud de nuevo con el nuevo token
-        const retryResponse = await fetch(`${BASE_URL}/api/exercise/delete-exercise/${exercise_id}`, {
-        method: 'DELETE',
-        headers: {
-            'Authorization': `Bearer ${newToken}`,
-        },
+        try {
+        const response = await fetch(`${BASE_URL}/api/exercise/delete-exercise/${exercise_id}`, {
+            method: 'DELETE',
+            headers: {
+            'Authorization': token,
+            },
         });
 
-        if (!retryResponse.ok) {
-        const errorData = await retryResponse.json();
-        throw new Error(errorData.error || 'Error al eliminar ejercicio');
+        // Si la respuesta es 401, intentamos renovar el token
+        if (response.status === 403 || response.status === 401) {
+            console.log('Token expirado, intentando renovar...');
+            const newToken = await refreshAuthToken(); // Renueva el token
+            // Intentamos la solicitud de nuevo con el nuevo token
+            const retryResponse = await fetch(`${BASE_URL}/api/exercise/delete-exercise/${exercise_id}`, {
+            method: 'DELETE',
+            headers: {
+                'Authorization': `Bearer ${newToken}`,
+            },
+            });
+
+            if (!retryResponse.ok) {
+            const errorData = await retryResponse.json();
+            throw new Error(errorData.error || 'Error al eliminar ejercicio');
+            }
+
+            const success = await retryResponse.json();
+            return success;
         }
 
-        const success = await retryResponse.json();
+        if (!response.ok) {
+            const errorData = await response.json();
+            throw new Error(errorData.error || 'Error al eliminar ejercicio');
+        }
+
+        const success = await response.json();
         return success;
-    }
-
-    if (!response.ok) {
-        const errorData = await response.json();
-        throw new Error(errorData.error || 'Error al eliminar ejercicio');
-    }
-
-    const success = await response.json();
-    return success;
-    } catch (error) {
-    console.error('Error al eliminar ejercicio:', error);
-    throw error;
-    }
+        } catch (error) {
+        console.error('Error al eliminar ejercicio:', error);
+        throw error;
+        }
+      }
 }
